Remove unused code found via staticcheck tool
