+++ agent/stats/utils.go
@@ -15,41 +15,16 @@ package stats
 
 import (
 	"math"
-	"regexp"
 	"runtime"
-	"time"
 
-	"github.com/cihub/seelog"
 	"github.com/docker/docker/api/types"
 )
-
-// networkStatsErrorPattern defines the pattern that is used to evaluate
-// if there's an error reading network stats.
-const networkStatsErrorPattern = "open /sys/class/net/veth.*: no such file or directory"
 
 var numCores = uint64(runtime.NumCPU())
 
 // nan32 returns a 32bit NaN.
 func nan32() float32 {
 	return (float32)(math.NaN())
-}
-
-// parseNanoTime returns the time object from a string formatted with RFC3339Nano layout.
-func parseNanoTime(value string) time.Time {
-	ts, _ := time.Parse(time.RFC3339Nano, value)
-	return ts
-}
-
-// isNetworkStatsError returns if the error indicates that files in /sys/class/net
-// could not be opened.
-func isNetworkStatsError(err error) bool {
-	matched, mErr := regexp.MatchString(networkStatsErrorPattern, err.Error())
-	if mErr != nil {
-		seelog.Debugf("Error matching string: %v", mErr)
-		return false
-	}
-
-	return matched
 }
 
 func getNetworkStats(dockerStats *types.StatsJSON) *NetworkStats {
